Unifies BHP commitments
