+++ algorithms/src/commitment/bhp.rs
@@ -33,7 +33,7 @@ pub struct BHPCommitment<G: ProjectiveCurve, const NUM_WINDOWS: usize, const WIN
 impl<G: ProjectiveCurve, const NUM_WINDOWS: usize, const WINDOW_SIZE: usize> CommitmentScheme
     for BHPCommitment<G, NUM_WINDOWS, WINDOW_SIZE>
 {
-    type Output = G::Affine;
+    type Output = <G::Affine as AffineCurve>::BaseField;
     type Parameters = (Vec<Vec<G>>, Vec<G>);
     type Randomness = G::ScalarField;
 
@@ -77,7 +77,9 @@ impl<G: ProjectiveCurve, const NUM_WINDOWS: usize, const WINDOW_SIZE: usize> Com
             }
         }
 
-        Ok(output.into_affine())
+        let affine = output.into_affine();
+        debug_assert!(affine.is_in_correct_subgroup_assuming_on_curve());
+        Ok(affine.to_x_coordinate())
     }
 
     fn parameters(&self) -> Self::Parameters {
