--- algorithms/src/commitment/bhp_compressed.rs
@@ -1,94 +1,0 @@
-// Copyright (C) 2019-2021 Aleo Systems Inc.
-// This file is part of the snarkVM library.
-
-// The snarkVM library is free software: you can redistribute it and/or modify
-// it under the terms of the GNU General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-
-// The snarkVM library is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-// GNU General Public License for more details.
-
-// You should have received a copy of the GNU General Public License
-// along with the snarkVM library. If not, see <https://www.gnu.org/licenses/>.
-
-use crate::{commitment::BHPCommitment, CommitmentError, CommitmentScheme};
-use snarkvm_curves::{AffineCurve, ProjectiveCurve};
-use snarkvm_fields::{ConstraintFieldError, Field, ToConstraintField};
-use snarkvm_utilities::{FromBytes, ToBytes};
-
-use std::{
-    fmt::Debug,
-    io::{Read, Result as IoResult, Write},
-};
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct BHPCompressedCommitment<G: ProjectiveCurve, const NUM_WINDOWS: usize, const WINDOW_SIZE: usize> {
-    pub bhp: BHPCommitment<G, NUM_WINDOWS, WINDOW_SIZE>,
-}
-
-impl<G: ProjectiveCurve, const NUM_WINDOWS: usize, const WINDOW_SIZE: usize> CommitmentScheme
-    for BHPCompressedCommitment<G, NUM_WINDOWS, WINDOW_SIZE>
-{
-    type Output = <G::Affine as AffineCurve>::BaseField;
-    type Parameters = BHPCommitment<G, NUM_WINDOWS, WINDOW_SIZE>;
-    type Randomness = G::ScalarField;
-
-    fn setup(message: &str) -> Self {
-        BHPCommitment::<G, NUM_WINDOWS, WINDOW_SIZE>::setup(message).into()
-    }
-
-    fn commit(&self, input: &[u8], randomness: &Self::Randomness) -> Result<Self::Output, CommitmentError> {
-        let affine = self.bhp.commit(input, randomness)?;
-        debug_assert!(affine.is_in_correct_subgroup_assuming_on_curve());
-        Ok(affine.to_x_coordinate())
-    }
-
-    fn parameters(&self) -> Self::Parameters {
-        self.bhp.clone()
-    }
-}
-
-impl<G: ProjectiveCurve, const NUM_WINDOWS: usize, const WINDOW_SIZE: usize>
-    From<BHPCommitment<G, NUM_WINDOWS, WINDOW_SIZE>> for BHPCompressedCommitment<G, NUM_WINDOWS, WINDOW_SIZE>
-{
-    fn from(bhp: BHPCommitment<G, NUM_WINDOWS, WINDOW_SIZE>) -> Self {
-        Self { bhp }
-    }
-}
-
-impl<G: ProjectiveCurve, const NUM_WINDOWS: usize, const WINDOW_SIZE: usize> From<(Vec<Vec<G>>, Vec<G>)>
-    for BHPCompressedCommitment<G, NUM_WINDOWS, WINDOW_SIZE>
-{
-    fn from((bases, random_base): (Vec<Vec<G>>, Vec<G>)) -> Self {
-        BHPCommitment::from((bases, random_base)).into()
-    }
-}
-
-impl<G: ProjectiveCurve, const NUM_WINDOWS: usize, const WINDOW_SIZE: usize> ToBytes
-    for BHPCompressedCommitment<G, NUM_WINDOWS, WINDOW_SIZE>
-{
-    fn write_le<W: Write>(&self, writer: W) -> IoResult<()> {
-        self.bhp.write_le(writer)
-    }
-}
-
-impl<G: ProjectiveCurve, const NUM_WINDOWS: usize, const WINDOW_SIZE: usize> FromBytes
-    for BHPCompressedCommitment<G, NUM_WINDOWS, WINDOW_SIZE>
-{
-    #[inline]
-    fn read_le<R: Read>(reader: R) -> IoResult<Self> {
-        Ok(BHPCommitment::read_le(reader)?.into())
-    }
-}
-
-impl<F: Field, G: ProjectiveCurve + ToConstraintField<F>, const NUM_WINDOWS: usize, const WINDOW_SIZE: usize>
-    ToConstraintField<F> for BHPCompressedCommitment<G, NUM_WINDOWS, WINDOW_SIZE>
-{
-    #[inline]
-    fn to_field_elements(&self) -> Result<Vec<F>, ConstraintFieldError> {
-        self.bhp.to_field_elements()
-    }
-}
